+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,7 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+- ui: add “notify.js”
 
 0.14.0
 ======